Refactor: Shop데이터를 fetch하는 코드 파일 분리
